build backend

+++ clubs/views.py
@@ -2,12 +2,15 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework.status import HTTP_404_NOT_FOUND, HTTP_201_CREATED, HTTP_422_UNPROCESSABLE_ENTITY, HTTP_204_NO_CONTENT, HTTP_202_ACCEPTED, HTTP_401_UNAUTHORIZED
+from rest_framework.permissions import IsAuthenticatedOrReadOnly
 
 from .models import Club
 from .serializers import PopulatedClubSerializer, ClubSerializer, UserSerializer, BookSerializer
 
 
 class ClubListView(APIView):
+
+  permission_classes = (IsAuthenticatedOrReadOnly, )
 
   def get(self, _request):
 
@@ -17,9 +20,10 @@ class ClubListView(APIView):
     return Response(serialized_clubs.data)
 
   def post(self, request):
-
+    request.data['owner'] = request.user.id
+    # club = Club.objects.get()   KEEP AN EYE ON THIS. NOT SURE WE need it
     club = ClubSerializer(data=request.data)
-
+    
     if club.is_valid():
       club.save()
       return Response(club.data, status=HTTP_201_CREATED)
@@ -28,6 +32,9 @@ class ClubListView(APIView):
 
 
 class ClubDetailView(APIView):
+
+  permission_classes = (IsAuthenticatedOrReadOnly, )
+
   def get(self, _request, pk):
     try:
       club = Club.objects.get(pk=pk)
@@ -38,9 +45,12 @@ class ClubDetailView(APIView):
 
   
   def put(self, request, pk):
+    request.data['owner'] = request.user.id
 
     try:
       club = Club.objects.get(pk=pk)
+      if club.owner.id != request.user.id:
+        return Response(status=HTTP_401_UNAUTHORIZED)
       updated_club = ClubSerializer(club, data=request.data)
       if updated_club.is_valid():
         updated_club.save()
@@ -49,9 +59,12 @@ class ClubDetailView(APIView):
     except Club.DoesNotExist:
       return Response({ 'message': 'Not Found'}, status=HTTP_404_NOT_FOUND)
 
-  def delete(self, _request, pk):
+  def delete(self, request, pk):
+    # request.data['owner'] = request.user.id
     try:
         club = Club.objects.get(pk=pk)
+        if club.owner.id != request.user.id:
+          return Response(status=HTTP_401_UNAUTHORIZED)
         club.delete()
         return Response(status=HTTP_204_NO_CONTENT)
     except Club.DoesNotExist:
